requirements.txt
kafka-python==2.0.2
pandas==2.2.0
scikit-learn==1.4.0
numpy==1.26.0
